+++ pkg/contracts/solana/instruction.go
@@ -1,8 +1,10 @@
 package solana
 
 import (
+	"encoding/json"
 	"fmt"
 
+	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/gagliardetto/solana-go"
@@ -142,7 +144,73 @@ func ParseInstructionWithdraw(instruction solana.CompiledInstruction) (*Withdraw
 
 	// check the discriminator to ensure it's a 'withdraw' instruction
 	if inst.Discriminator != DiscriminatorWithdraw {
-		return nil, fmt.Errorf("not a withdraw instruction: %v", inst.Discriminator)
+		return nil, fmt.Errorf("not a withdraw instruction: %v %v", inst.Discriminator, DiscriminatorWithdraw)
+	}
+
+	return inst, nil
+}
+
+var _ OutboundInstruction = (*ExecuteInstructionParams)(nil)
+
+// ExecuteInstructionParams contains the parameters for a gateway execute instruction
+type ExecuteInstructionParams struct {
+	// Discriminator is the unique identifier for the execute instruction
+	Discriminator [8]byte
+
+	// Amount is the lamports amount for the execute
+	Amount uint64
+
+	// Sender from zetachain
+	Sender [20]byte
+
+	// Data for connected program
+	Data []byte
+
+	// Signature is the ECDSA signature (by TSS) for the execute
+	Signature [64]byte
+
+	// RecoveryID is the recovery ID used to recover the public key from ECDSA signature
+	RecoveryID uint8
+
+	// MessageHash is the hash of the message signed by TSS
+	MessageHash [32]byte
+
+	// Nonce is the nonce for the execute
+	Nonce uint64
+}
+
+// Signer returns the signer of the signature contained
+func (inst *ExecuteInstructionParams) Signer() (signer common.Address, err error) {
+	var signature [65]byte
+	copy(signature[:], inst.Signature[:64])
+	signature[64] = inst.RecoveryID
+
+	return RecoverSigner(inst.MessageHash[:], signature[:])
+}
+
+// GatewayNonce returns the nonce of the instruction
+func (inst *ExecuteInstructionParams) GatewayNonce() uint64 {
+	return inst.Nonce
+}
+
+// TokenAmount returns the amount of the instruction
+func (inst *ExecuteInstructionParams) TokenAmount() uint64 {
+	return inst.Amount
+}
+
+// ParseInstructionExecute tries to parse the instruction as a 'execute'.
+// It returns nil if the instruction can't be parsed as a 'execute'.
+func ParseInstructionExecute(instruction solana.CompiledInstruction) (*ExecuteInstructionParams, error) {
+	// try deserializing instruction as a 'execute'
+	inst := &ExecuteInstructionParams{}
+	err := borsh.Deserialize(inst, instruction.Data)
+	if err != nil {
+		return nil, errors.Wrap(err, "error deserializing instruction")
+	}
+
+	// check the discriminator to ensure it's a 'execute' instruction
+	if inst.Discriminator != DiscriminatorExecute {
+		return nil, fmt.Errorf("not an execute instruction: %v", inst.Discriminator)
 	}
 
 	return inst, nil
@@ -274,4 +342,66 @@ func ParseInstructionWhitelist(instruction solana.CompiledInstruction) (*Whiteli
 	}
 
 	return inst, nil
+}
+
+type AccountMeta struct {
+	PublicKey  [32]byte
+	IsWritable bool
+}
+
+// ExecuteMsg describes data and accounts passed to connected programs
+type ExecuteMsg struct {
+	Accounts []AccountMeta
+	Data     []byte
+}
+
+// GetExecuteMsgAbi used for abi encoding/decoding execute msg
+func GetExecuteMsgAbi() (abi.Arguments, error) {
+	MsgAbiType, err := abi.NewType("tuple", "struct Msg", []abi.ArgumentMarshaling{
+		{
+			Name: "accounts",
+			Type: "tuple[]",
+			Components: []abi.ArgumentMarshaling{
+				{Name: "publicKey", Type: "bytes32"},
+				{Name: "isWritable", Type: "bool"},
+			},
+		},
+		{Name: "data", Type: "bytes"},
+	})
+
+	if err != nil {
+		return abi.Arguments{}, err
+	}
+
+	MsgAbiArgs := abi.Arguments{
+		{Type: MsgAbiType, Name: "msg"},
+	}
+
+	return MsgAbiArgs, nil
+}
+
+// DecodeExecuteMsg decodes execute msg using abi decoding
+func DecodeExecuteMsg(msgbz []byte) (ExecuteMsg, error) {
+	args, err := GetExecuteMsgAbi()
+	if err != nil {
+		return ExecuteMsg{}, err
+	}
+
+	unpacked, err := args.Unpack(msgbz)
+	if err != nil {
+		return ExecuteMsg{}, err
+	}
+
+	jsonMsg, err := json.Marshal(unpacked[0])
+	if err != nil {
+		return ExecuteMsg{}, err
+	}
+
+	var msg ExecuteMsg
+	err = json.Unmarshal(jsonMsg, &msg)
+	if err != nil {
+		return ExecuteMsg{}, err
+	}
+
+	return msg, nil
 }