+++ x/crosschain/client/cli/tx.go
@@ -31,6 +31,7 @@ func GetTxCmd() *cobra.Command {
 		CmdWhitelistERC20(),
 		CmdAbortStuckCCTX(),
 		CmdRefundAborted(),
+		CmdRemoveInboundTracker(),
 	)
 
 	return cmd